Create a KV type that key-value sequence mappers use.
The mappers previously stored internally stored a Vec<String, String> (or
a u8 equivalent) and output [String, String]. This caused a mismatch
with other matchers since the output contained owned data (String)
rather than borrowed data (str) like the rest of the mappers. Mappers
can not specify an Out type of [&str, &str], because there is no place
to annotate the correct lifetime for the &str types. To address this the
KV type has two type parameters (K and V) and internally stores owned
versions of K and V. The tuple mapper is then updated to only accept KV
input.

These changes remove the need for the deref mapper so it's removed.