hyper needs to bind to a port within a tokio context.
This seems unfortunate and really highlights why I hate TLS variables so
much. This hidden tokio context is necessary even for functions that
don't exhibit any tokio behavior. Binding doesn't even involve a future
and yet it will fail with a tokio reactor not found if done outside of a
tokio context. Anyway the fix is straightforward. Bind within the future
and transfer the address out using a non-blocking channel.