Have Fn(&T) -> bool implement Mapper<T>
Since Fn traits do not implement Debug this meant removing the
fmt::Debug requirement from Mapper. To still allow the server to
log/print the mapper name each mapper now implements a fmt method to
describe itself, and there is a mapper_name convenience function for
formatting the name of the mapper.