Rewrite the request::body mapper.
It previously attempted to implement mapper for any http::Request with a
body that implemented ToOwned. Unfortunately, ToOwned of a &&str results
in an &str not a String. Since the body method of http::Request returns
a reference to the body any http::Request<&str> would return &&str, and
the common mappers are only implemented for &str input, not &&str input.
This had a similar failure for http::Request<&[u8]>.

Rather than let perfect be the enemy of good enough body is now
implemented for &str, &[u8], Vec<u8>, and hyper::body::Bytes. This
should be good enough for most users.