Simplify, removing unnecessary flexibility.
Remove the Out parameter from the Mapper trait. Everything just returned
bool anyway. Rename Mapper to Matcher since it's always a boolean, and
the method from map to matches. This simplification should make it
easier to be slightly more opinionated and lay the groundwork for better
debugging of why matchers do or do not match.