Ensure that the state of a server within a ServerPool is reset when the handle is dropped
There was a bug (reported as #19) where if a test was already panicking
when a server handle was being dropped the drop implementation would not
clear the state of the server. This lead to subsequent tests using the
same server sporadically failing because the expectations had been
carried over from a prior test.