Print expectation failures to stdout when the server is dropped due to unwinding a panic
Prior to this change the server would not validate expectation if the test
panic's for reasons unrelated to the server expecations failing to be met.
However, this could make it difficult to find the issue if the test panic was
the result of a misconfiguration of the expectations. With this change the
server will now validate the server expectations and on failure will simply
print the failure message to stdout.